// apps/flow/lib/zoho/types.ts
export interface ZohoVendor {
  vendor_id: string
  contact_name: string
  company_name?: string
  email?: string
  phone?: string
  currency_code: string
  payment_terms: number
  payment_terms_label: string
  created_time: string
  last_modified_time: string
}